+++ front/src/app/feedlist/[word_id]/feedlist.module.scss
